+++ packages/create/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/create
 
+## 0.12.54
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.2.42
+-   @sap-ux/adp-tooling@0.13.31
+-   @sap-ux/flp-config-inquirer@0.2.76
+-   @sap-ux/preview-middleware@0.19.25
+-   @sap-ux/app-config-writer@0.5.46
+
 ## 0.12.53
 
 ### Patch Changes